.Core cosmetics
git-svn-id: svn://ultimatepp.org/upp/trunk@5083 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/Core/TimeDate.cpp
@@ -575,7 +575,7 @@ String GetTimeZoneText()
 {
 	Time gmtime = GetUtcTime();
 	Time ltime = GetSysTime();
-	int d = (ltime - gmtime) / 600;
+	int d = (int)(ltime - gmtime) / 600;
 	return Format("+%02.2d%01.1d0", d / 6, d % 6);
 }
